post_twitter/requirements.txt
functions-framework==3.4.0
google-cloud-firestore==2.11.1
tweepy==4.14.0
python-dotenv==1.0.0